Module3 Task 1
In the _1 file we follow the tutorial, and at the end we make the linear model for the cars example. The FixedScript is the exercise of fixing some lines of code where a linear regression model is constructed for the iris.csv. We include a PDF version with the printed report.